Adding 3 files relating user inputs to webpage for recipe search
IngredientInputPage.php is where most of the user input is happening then it goes to the other PHP files. The user can enter one ingredient at a time and see the item be added to the list. They can also delete a certain item in the list before sending it to the database. Once they are done adding to the list, there is an option to send the list to the database. The submit button does not work yet.  I also have the part where the user can upload a picture to the database under that as another option instead of typing your list. Sending a picture does work but it is set up to send it to my local database.